+++ Source/Nett.AspNet/TomlConfigurationProvider.cs
@@ -1,4 +1,5 @@
-﻿using System.IO;
+﻿using System;
+using System.IO;
 using Microsoft.Extensions.Configuration;
 using Nett.Extensions;
 
@@ -6,16 +7,28 @@ namespace Nett.AspNet
 {
     public sealed class TomlConfigurationProvider : FileConfigurationProvider
     {
+        public bool CaseInsensitiveKeys { get; }
+
         public TomlConfigurationProvider(TomlConfigurationSource source)
             : base(source)
         {
             source.CheckNotNull(nameof(source));
+
+            this.CaseInsensitiveKeys = source.CaseInsensitiveKeys;
         }
 
         public override void Load(Stream stream)
         {
             var table = Toml.ReadStream(stream);
-            this.Data = ProviderDictionaryConverter.ToProviderDictionary(table);
+
+            if (this.CaseInsensitiveKeys)
+            {
+                this.Data = ProviderDictionaryConverter.ToProviderDictionary(table, StringComparer.OrdinalIgnoreCase);
+            }
+            else
+            {
+                this.Data = ProviderDictionaryConverter.ToProviderDictionary(table);
+            }
         }
     }
 }