+++ Source/Nett.AspNet/TomlConfigurationSource.cs
@@ -4,6 +4,8 @@ namespace Nett.AspNet
 {
     public class TomlConfigurationSource : FileConfigurationSource
     {
+        public bool CaseInsensitiveKeys { get; set; }
+
         public override IConfigurationProvider Build(IConfigurationBuilder builder)
         {
             return new TomlConfigurationProvider(this);