+++ RELEASENOTES.md
@@ -2,6 +2,9 @@
 Nett:
 + Add: Callback for processing non map-able table rows [#78](https://github.com/paiden/Nett/issues/78)
 
+AspNet:
++ Add: Option to treat provider dictionary config keys as case insensitive [#81](https://github.com/paiden/Nett/issues/81)
+
 ## v0.12.0
 Nett:
 + Fix: ArgumentOutOfRange exception when DateTime field/prop not initialized [#25](https://github.com/paiden/Nett/issues/25)